+++ src/types/types.ts
@@ -1,0 +1,13 @@
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface Todo {
+  id: number;
+  title: string;
+  userId: number;
+  completed: boolean;
+  user: User;
+}
